# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel APIs Generator ",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get a listing of the Orders.",
                "description": "Get all Orders",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User Auth Token{ Bearer ABC123 }",
                        "required": true,
                        "type": "string",
                        "default": "Bearer ABC123"
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Pass the property name you want to sort your response. If not found, Returns All Records in DB without sorting.",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "sortedBy",
                        "in": "query",
                        "description": "Pass 'asc' or 'desc' to define the sorting method. If not found, 'asc' will be used by default",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Change the Default Record Count. If not found, Returns All Records in DB.",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Change the Default Offset of the Query. If not found, 0 will be used.",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Order"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Store a newly created Order in storage",
                "description": "Store Order",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User Auth Token{ Bearer ABC123 }",
                        "required": true,
                        "type": "string",
                        "default": "Bearer ABC123"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Order that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Order"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Display the specified Order",
                "description": "Get Order",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User Auth Token{ Bearer ABC123 }",
                        "required": true,
                        "type": "string",
                        "default": "Bearer ABC123"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Order",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Order"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Order"
                ],
                "summary": "Update the specified Order in storage",
                "description": "Update Order",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User Auth Token{ Bearer ABC123 }",
                        "required": true,
                        "type": "string",
                        "default": "Bearer ABC123"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Order",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Order that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Order"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Order"
                ],
                "summary": "Remove the specified Order from storage",
                "description": "Delete Order",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User Auth Token{ Bearer ABC123 }",
                        "required": true,
                        "type": "string",
                        "default": "Bearer ABC123"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Order",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/cancel-order": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "cancel a newly created Order in storage",
                "description": "Store Order",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User Auth Token{ Bearer ABC123 }",
                        "required": true,
                        "type": "string",
                        "default": "Bearer ABC123"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Order that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Order"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/orders/rating": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Store a newly created Order in storage",
                "description": "Store Order",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "User Auth Token{ Bearer ABC123 }",
                        "required": true,
                        "type": "string",
                        "default": "Bearer ABC123"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Order that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Order"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Order": {
            "required": [
                "user_id",
                "cart_id",
                "petspace_id",
                "user_address_id",
                "technician_id",
                "slot_id",
                "date_time",
                "status",
                "progress_status",
                "sub_total",
                "tax",
                "delivery_fee",
                "total",
                "rating",
                "rating_comment",
                "note",
                "promo_code"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "user_id": {
                    "description": "user_id",
                    "type": "integer",
                    "format": "int32"
                },
                "petspace_id": {
                    "description": "petspace_id",
                    "type": "integer",
                    "format": "int32"
                },
                "user_address_id": {
                    "description": "user_address_id",
                    "type": "integer",
                    "format": "int32"
                },
                "technician_id": {
                    "description": "technician_id",
                    "type": "integer",
                    "format": "int32"
                },
                "slot_id": {
                    "description": "slot_id",
                    "type": "integer",
                    "format": "int32"
                },
                "status": {
                    "description": "status",
                    "type": "integer",
                    "format": "int32"
                },
                "date_time": {
                    "description": "date_time",
                    "type": "string",
                    "format": "date-time"
                },
                "sub_total": {
                    "description": "sub_total",
                    "type": "number",
                    "format": "number"
                },
                "delivery_fee": {
                    "description": "delivery_fee",
                    "type": "number",
                    "format": "number"
                },
                "tax": {
                    "description": "tax",
                    "type": "number",
                    "format": "number"
                },
                "total": {
                    "description": "total",
                    "type": "number",
                    "format": "number"
                },
                "note": {
                    "description": "note",
                    "type": "string"
                },
                "promo_code": {
                    "description": "promo_code",
                    "type": "string"
                }
            }
        },
...
    }
}